1560: Backport command fails to load census
