+++ bots/pr/src/main/java/org/openjdk/skara/bots/pr/TagCommand.java
@@ -59,7 +59,8 @@ public class TagCommand implements CommandHandler {
     }
 
     @Override
-    public void handle(PullRequestBot bot, HostedCommit commit, CensusInstance censusInstance, Path scratchPath, CommandInvocation command, List<Comment> allComments, PrintWriter reply) {
+    public void handle(PullRequestBot bot, HostedCommit commit, LimitedCensusInstance censusInstance,
+            Path scratchPath, CommandInvocation command, List<Comment> allComments, PrintWriter reply) {
         try {
             if (censusInstance.contributor(command.user()).isEmpty()) {
                 reply.println("Only OpenJDK [contributors](https://openjdk.org/bylaws#contributor) can use the `/tag` command.");
